ARM: also set GPIO function and mode.
This also implements more of the FastIO infrastructure.
Unfortunately, all these registers are located and used in a not
very aligned manner, so we need lots of tabular data. Headers
coming with MBED contain not much useful stuff either, so I had
to step through the user manual, pin by pin.