Messages: more newlines needed.
The recent switch to send 'ok' postponed requires also sending a
newline in a few places, because this 'ok' is no longer at the
start of the line. Now it appears in its own line.

A recent commit tried to correct this already, but it missed a
few locations.  Also remove the comments promising a forthcoming
newline from gcode_parse.