Isolate Teacup source code to appease Arduino IDE
Recent arduino IDEs (1.6 and later?) have become more aggressive
about searching for source code to compile recursively.  This breaks
IDE builds when the compile tries to compile old "attic" and "extruder"
code, and also the "simulator" code which is already in a subdirectory
specifically to hide it from the IDE.

Move the Arduino source into a properly named subdirectory to make
the IDE happy and adapt the makefiles to find it.

TODO: Update the wiki to teach users how to launch the project in the IDE.
TODO: Update configtool to use the new paths.